Just move where the crash rollup functionality happens
