+++ src/data/join_segments_crash.py
@@ -10,6 +10,10 @@ import json
 from . import util
 import os
 import argparse
+from pandas.io.json import json_normalize
+from shapely.geometry import Point
+import geopandas as gpd
+
 
 BASE_DIR = os.path.dirname(
     os.path.dirname(
@@ -42,6 +46,59 @@ def snap_records(
         json.dump([r.properties for r in records], f)
 
 
+def calculate_crashes_by_location(df):
+    """
+    Calculates total number of crashes that occurred at each unique lat/lng pair and
+    generates a comma-separated string of the dates that crashes occurred at that location
+
+    Inputs:
+        - a dataframe where each row represents one unique crash incident
+
+    Output:
+        - a dataframe with the total number of crashes at each unique crash location
+          and list of unique crash dates
+    """
+    crashes_agg = df.groupby(['latitude', 'longitude']).agg(['count', 'unique'])
+    crashes_agg.columns = crashes_agg.columns.get_level_values(1)
+    crashes_agg.rename(columns={'count': 'total_crashes', 'unique': 'crash_dates'}, inplace=True)
+    crashes_agg.reset_index(inplace=True)
+    
+    crashes_agg['crash_dates'] = crashes_agg['crash_dates'].str.join(',')
+    return crashes_agg
+
+
+def make_crash_rollup(crashes_json):
+    """
+    Generates a GeoDataframe with the total number of crashes and a comma-separated string
+    of crash dates per unique lat/lng pair
+
+    Inputs:
+        - a json of standardized crash data
+
+    Output:
+        - a GeoDataframe with the following columns:
+            - total number of crashes
+            - list of unique dates that crashes occurred
+            - GeoJSON point features created from the latitude and longitude
+    """
+    df_std_crashes = json_normalize(crashes_json)
+    df_std_crashes = df_std_crashes[
+        ["dateOccurred", "location.latitude", "location.longitude"]]
+    df_std_crashes.rename(columns={
+        "location.latitude": "latitude",
+        "location.longitude": "longitude"
+    }, inplace=True)
+
+    crashes_agg = calculate_crashes_by_location(df_std_crashes)
+    crashes_agg["coordinates"] = list(zip(
+        crashes_agg.longitude, crashes_agg.latitude))
+    crashes_agg["coordinates"] = crashes_agg["coordinates"].apply(Point)
+    crashes_agg = crashes_agg[["coordinates", "total_crashes", "crash_dates"]]
+
+    crashes_agg_gdf = gpd.GeoDataFrame(crashes_agg, geometry="coordinates")
+    return crashes_agg_gdf
+
+
 if __name__ == '__main__':
 
     parser = argparse.ArgumentParser()
@@ -66,3 +123,15 @@ if __name__ == '__main__':
         os.path.join(RAW_DATA_FP, 'crashes.json'),
         startyear=args.startyear, endyear=args.endyear)
 
+    with open(os.path.join(RAW_DATA_FP, 'crashes.json')) as crash_file:
+        data = json.load(crash_file)
+    crashes_agg_gdf = make_crash_rollup(data)
+
+    crashes_agg_path = os.path.join(
+        args.datadir, "processed", "crashes_rollup.geojson")
+    if os.path.exists(crashes_agg_path):
+        os.remove(crashes_agg_path)
+    crashes_agg_gdf.to_file(
+        os.path.join(args.datadir, "processed", "crashes_rollup.geojson"),
+        driver="GeoJSON"
+    )
